backend/pyproject.toml
[project]
name = "drf-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "drf-nested-routers>=0.94.2",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "django-browser-reload>=1.18.0",
    "django-debug-toolbar>=6.0.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.8",
]
